@@ -8,7 +8,7 @@ class CardCategory extends StatelessWidget {
       this.tap = defaultFunc});
 
   final String img;
-  final Function tap;
+  final VoidCallback tap;
   final String title;
 
   static void defaultFunc() {
@@ -31,7 +31,10 @@ class CardCategory extends StatelessWidget {
                     decoration: BoxDecoration(
                         borderRadius: BorderRadius.all(Radius.circular(6.0)),
                         image: DecorationImage(
-                          image: NetworkImage(img),
+                          // Aquí chequeamos si es un asset local o una URL
+                          image: img.startsWith('assets/')
+                              ? AssetImage(img)
+                              : NetworkImage(img) as ImageProvider<Object>,
                           fit: BoxFit.cover,
                         ))),
                 Container(
