@@ -10,7 +10,7 @@ class CardHorizontal extends StatelessWidget {
 
   final String cta;
   final String img;
-  final Function tap;
+  final VoidCallback tap;
   final String title;
 
   static void defaultFunc() {
@@ -38,7 +38,10 @@ class CardHorizontal extends StatelessWidget {
                               topLeft: Radius.circular(4.0),
                               bottomLeft: Radius.circular(4.0)),
                           image: DecorationImage(
-                            image: NetworkImage(img),
+                            // Aquí chequeamos si es un asset local o una URL
+                            image: img.startsWith('assets/')
+                                ? AssetImage(img)
+                                : NetworkImage(img) as ImageProvider<Object>,
                             fit: BoxFit.cover,
                           ))),
                 ),
