@@ -2,15 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:now_ui_flutter/constants/Theme.dart';
 
 class CardSquare extends StatelessWidget {
-  CardSquare(
-      {this.title = "Placeholder Title",
-      this.cta = "",
-      this.img = "https://via.placeholder.com/200",
-      this.tap = defaultFunc});
+  CardSquare({
+    this.title = "Placeholder Title",
+    this.cta = "",
+    this.img = "https://via.placeholder.com/200",
+    required this.tap,
+  });
 
   final String cta;
   final String img;
-  final Function tap;
+  final VoidCallback tap;
   final String title;
 
   static void defaultFunc() {
@@ -40,7 +41,11 @@ class CardSquare extends StatelessWidget {
                                   topLeft: Radius.circular(4.0),
                                   topRight: Radius.circular(4.0)),
                               image: DecorationImage(
-                                image: NetworkImage(img),
+                                // Aquí chequeamos si es un asset local o una URL
+                                image: img.startsWith('assets/')
+                                    ? AssetImage(img)
+                                    : NetworkImage(img)
+                                        as ImageProvider<Object>,
                                 fit: BoxFit.cover,
                               )))),
                   Flexible(
